2024/Web/QrZilla/src/app/requirements.txt
Flask==2.0.2
PyMySQL==1.0.2
pyqrcode==1.2.1
Jinja2==3.0.3
Pillow==9.0.1
Werkzeug==2.0.2
requests==2.26.0
pyzbar
pypng